List models (#123)
* List models

+++ management/management_api/models/model_utils.py
@@ -1,11 +1,44 @@
 import argparse
 from kubernetes.client.rest import ApiException
+from botocore.exceptions import ClientError
 
 from management_api.config import minio_resource
 from management_api.utils.kubernetes_resources import get_k8s_apps_api_client
 from management_api.utils.errors_handling import KubernetesGetException, ModelDeleteException, \
-    ModelDoesNotExistException, TenantDoesNotExistException
+    ModelDoesNotExistException, TenantDoesNotExistException, MinioCallException
 from management_api.tenants.tenants_utils import tenant_exists
+
+
+def list_models(namespace: str):
+    if not tenant_exists(namespace):
+        raise TenantDoesNotExistException(namespace)
+
+    apps_api_client = get_k8s_apps_api_client()
+    try:
+        deployments = apps_api_client.list_namespaced_deployment(namespace)
+    except ApiException as apiException:
+        raise KubernetesGetException('deployment', apiException)
+
+    try:
+        bucket = minio_resource.Bucket(name=namespace)
+    except ClientError as clientError:
+        raise MinioCallException(f'An error occurred during bucket reading: {clientError}')
+
+    models = []
+    for object in bucket.objects.all():
+        if object.size > 0:
+            model_path = object.key.split('/')[0].rsplit('-', 1)
+            model_name = model_path[0]
+            model_version = model_path[-1]
+            model_size = object.size
+            deployment_count = len(endpoints_using_model(deployments, model_path))
+            models.append((model_name, model_version, model_size, deployment_count))
+
+    if not models:
+        return f"There are no models present in {namespace} tenant\n"
+    else:
+        return f'Models in {namespace} tenant ' \
+               f'(model name, model version, model size, deployed count): {models}\n'
 
 
 def delete_model(parameters: dict, namespace: str):
@@ -25,7 +58,9 @@ def delete_model(parameters: dict, namespace: str):
     except ApiException as apiException:
         raise KubernetesGetException('deployment', apiException)
 
-    endpoints_using_model(deployments, model_path)
+    endpoint_names = endpoints_using_model(deployments, model_path)
+    if endpoint_names:
+        raise ModelDeleteException(f'model is used by endpoints: {endpoint_names}')
 
     for key in model_in_bucket:
         key.delete()
@@ -41,8 +76,8 @@ def endpoints_using_model(deployments, model_path):
         model_base_path = get_model_base_path(arg)
         if model_path == model_base_path:
             endpoint_names.append(endpoint_name)
-    if endpoint_names:
-        raise ModelDeleteException(f'model is used by endpoints: {endpoint_names}')
+
+    return endpoint_names
 
 
 def model_exists(model):
